ipa: Use local DATA_IPACFG_MGR.
Since #237825 lineage provides a device independend implementation
of ipa, based on sdm845.
This is incompatible to our kernel version and we are using an own
copy of the ipa implementation. As long as we don't adapt our kernel
ipa implementation to match the sdm845 one, we use this copy for now.

Change-Id: Ide07958fc17afe415aa0f3d5b496c8a0371c83d0